Add output

+++ src/main/java/org/quydusaigon/predatorsim/util/PreyStat.java
@@ -3,10 +3,10 @@ package org.quydusaigon.predatorsim.util;
 public class PreyStat extends AnimalStat {
 
     public PreySize size;
-    public int nutrition;
+    public Double nutrition;
     public double defense;
 
-    public PreyStat(double runSpeed, double visionRange, PreySize size, int nutrition, double defense) {
+    public PreyStat(double runSpeed, double visionRange, PreySize size, double nutrition, double defense) {
         super(runSpeed, visionRange);
         this.size = size;
         this.nutrition = nutrition;
